chore: run forge fmt

+++ contracts/common/lib/ECVerify.sol
@@ -1,12 +1,7 @@
 pragma solidity ^0.5.2;
 
-
 library ECVerify {
-    function ecrecovery(bytes32 hash, uint[3] memory sig)
-        internal
-        pure
-        returns (address)
-    {
+    function ecrecovery(bytes32 hash, uint256[3] memory sig) internal pure returns (address) {
         bytes32 r;
         bytes32 s;
         uint8 v;
@@ -39,11 +34,7 @@ library ECVerify {
         return result;
     }
 
-    function ecrecovery(bytes32 hash, bytes memory sig)
-        internal
-        pure
-        returns (address)
-    {
+    function ecrecovery(bytes32 hash, bytes memory sig) internal pure returns (address) {
         bytes32 r;
         bytes32 s;
         uint8 v;
@@ -76,11 +67,7 @@ library ECVerify {
         return result;
     }
 
-    function ecrecovery(bytes32 hash, uint8 v, bytes32 r, bytes32 s)
-        internal
-        pure
-        returns (address)
-    {
+    function ecrecovery(bytes32 hash, uint8 v, bytes32 r, bytes32 s) internal pure returns (address) {
         // get address out of hash and signature
         address result = ecrecover(hash, v, r, s);
 
@@ -90,11 +77,7 @@ library ECVerify {
         return result;
     }
 
-    function ecverify(bytes32 hash, bytes memory sig, address signer)
-        internal
-        pure
-        returns (bool)
-    {
+    function ecverify(bytes32 hash, bytes memory sig, address signer) internal pure returns (bool) {
         return signer == ecrecovery(hash, sig);
     }
 }
